separate the testing for file uploader into unit and integration tests
